Fixed major bug and refactored
Originally used Vec<<Mutex<dyn CommandHandler>> to silence the compiler since I am very new to rust but of course that had the complete wrong effect on the behavior of CommandHander. It was sharing state between all other command handlers of the same kind.

Refactored to use Box<dyn CommandHandler> and implemented the box copy handling (Which was the right way to implement this).